third_party/WebKit/LayoutTests/virtual/threaded/fast/animationworklet/worklet-animation-currentTime.html

<!DOCTYPE html>

<script id="simple_animate" type="text/worklet">
registerAnimator("test_animator", class {
  animate(currentTime, effect) {
    if (this.firstCurrentTime == undefined) {
      this.firstCurrentTime = currentTime;
      console.log(`first currentTime is ${this.firstCurrentTime}`);
    } else if (this.secondCurrentTime == undefined) {
      this.secondCurrentTime = currentTime;
      console.log(`second currentTime > first currentTime is ${this.secondCurrentTime > this.firstCurrentTime}`)
    }
  }
});
</script>
<div id="target"></div>
<script src="resources/animation-worklet-tests.js"></script>
<script>
if (window.testRunner) {
  testRunner.waitUntilDone();
  testRunner.dumpAsText();
}

runInAnimationWorklet(
  document.getElementById('simple_animate').textContent
).then(_ => {
  const effect = new KeyframeEffect(document.getElementById("target"), [{ opacity: 0 }], { duration: 1000 });
  const animation = new WorkletAnimation('test_animator', effect);
  animation.play();

  if (window.testRunner) {
    waitTwoAnimationFrames( _ => {
      waitTwoAnimationFrames( _ => {
        testRunner.notifyDone();
      });
    });
  }
});
</script>
