third_party/WebKit/LayoutTests/virtual/threaded/fast/animationworklet/worklet-animation-currentTime-expected.txt

CONSOLE MESSAGE: line 6: first currentTime is 0
CONSOLE MESSAGE: line 9: second currentTime > first currentTime is true
